FogBugz #224904 - Add systemd watchdog off during shutdown.
It was reported, and observed during QSPI boot testing, that reboot on
Angstrom file system caused an unclean shutdown.  Debugging showed
that systemd was attempting to use the watchog during shutdown, but
this interfered with the watchdog daemon, causing the watchdog to fire,
and thereby rebooting the system before shutdown was complete.

The solution was to tell systemd not to use the watchdog during shutdown.
A complete system.conf file was added, via bbappand.

Signed-off-by: Graham Moore <grmoore@opensource.altera.com>